requirements.txt
streamlit
llama-index-core
llama-index-indices-managed-llama-cloud
google-genai
llama-index-llms-google-genai